chore: code refinement
